Ruby 2.5 compatibility
We got errors where the lib tried to do BigDecimal(""), which raises an
error in Ruby 2.5 but returned 0 (or nil?) in Ruby 2.2.

I believe the culprit is another part of this lib that runs code like

     Common::Amount.new(
       :total => attributes[:amount],
       :tax => attributes[:tax_amount],
       :shipping => attributes[:shipping_amount]
     )

where I believe the tax_amount and shipping_amount are nil, causing this issue.